// src/types/register/regIsterErrors.ts
export enum REGISTER_ERROR_TEXTS {
    NAME_EMPTY = "Name is empty",
    EMAIL_EMPTY = "Email is empty",
    EMAIL_INVALID = "Invalid Email",
    PASSWORD_EMPTY = "Password is empty",
    INVALID_PHOTO_URL= "Invalid Image URL",
    PHONE_EMPTY = "phone empty",
    INVALID_PHONE= "Invalid phone",
    EMPTY_FIELD = "Empty & required field, please fill"
}